+++ app/src/main/res/layout/fragment_home.xml
@@ -38,7 +38,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/tv_home_greeting" />
 
-        <com.example.ishopping.ui.SearchHomeLayout
+        <com.example.ishopping.ui.search.SearchHomeLayout
             android:id="@+id/search_home_layout"
             android:layout_width="0dp"
             android:layout_height="58dp"
